Refactor import service utilities into modules
