+++ utils/custom_xml.py
@@ -1,0 +1,100 @@
+from __future__ import annotations
+
+import xml.etree.ElementTree as ET
+import zipfile
+from typing import Dict, List, Optional
+
+from config.settings import DEBUG_PRINT
+from utils.helpers import _normalize
+
+
+def _strip_xml_tag(tag: str) -> str:
+    """Remove namespace from an XML tag."""
+
+    return tag.split("}", 1)[1] if "}" in tag else tag
+
+
+def _element_to_flat_dict(elem: ET.Element, prefix: str = "") -> Dict[str, str]:
+    """
+    Recursively flatten an XML element into a key-value dict.
+    Nested elements become keys joined with underscores.
+
+    Example:
+        <participant>
+            <name>John</name>
+            <organization>MOI</organization>
+        </participant>
+        → {"participant_name": "John", "participant_organization": "MOI"}
+    """
+
+    children = list(elem)
+    if not children:
+        key = prefix or _strip_xml_tag(elem.tag)
+        return {key: (elem.text or "").strip()}
+
+    data: Dict[str, str] = {}
+    for child in children:
+        child_tag = _strip_xml_tag(child.tag)
+        child_prefix = f"{prefix}_{child_tag}" if prefix else child_tag
+        child_data = _element_to_flat_dict(child, child_prefix)
+        for key, value in child_data.items():
+            if not value:
+                continue
+            if key in data and data[key]:
+                data[key] = f"{data[key]}; {value}"
+            else:
+                data[key] = value
+    return data
+
+
+def _collect_custom_xml_records(path: str) -> Optional[Dict[str, List[Dict[str, str]]]]:
+    """
+    Collect embedded CustomXML parts from an Excel .xlsx file.
+
+    Returns:
+        dict mapping category → list[record_dict] or None if no records.
+        Example keys: 'participant', 'event', 'participant_event'.
+    """
+
+    try:
+        with zipfile.ZipFile(path) as zf:
+            names = [
+                n for n in zf.namelist()
+                if n.startswith("customXml/") and n.endswith(".xml")
+            ]
+            if not names:
+                return None
+
+            collected = {
+                "participant": [],
+                "event": [],
+                "participant_event": [],
+            }
+
+            for name in names:
+                try:
+                    root = ET.fromstring(zf.read(name))
+                except ET.ParseError:
+                    if DEBUG_PRINT:
+                        print(f"[CUSTOM-XML] Failed to parse {name}")
+                    continue
+
+                # Depth-first traversal
+                stack = [root]
+                while stack:
+                    node = stack.pop()
+                    tag = _strip_xml_tag(node.tag)
+                    if tag in collected:
+                        collected[tag].append(_element_to_flat_dict(node))
+                    stack.extend(list(node))
+
+            if not any(collected.values()):
+                return None
+
+            return {
+                "participants": collected["participant"],
+                "events": collected["event"],
+                "participant_events": collected["participant_event"],
+            }
+    except (zipfile.BadZipFile, FileNotFoundError):
+        return None
